move SVG stuff from theme to internal svg package
Thus, `ColorizeSVG` is available for other packages without being public.
