+++ internal/svg/testdata/colorized/circles.png
