+++ internal/svg/testdata/colorized/circles_white.png
