+++ internal/svg/testdata/colorized/circles_white_translucent.png
