+++ internal/svg/testdata/colorized/circles_yellow.png
