+++ internal/svg/testdata/colorized/circles_yellow_translucent.png
