+++ internal/svg/testdata/colorized/group_circles.png
