+++ internal/svg/testdata/colorized/group_paths.png
