+++ internal/svg/testdata/colorized/group_polygons.png
