+++ internal/svg/testdata/colorized/group_rects.png
