+++ internal/svg/testdata/colorized/paths.png
