+++ internal/svg/testdata/colorized/polygons.png
