+++ internal/svg/testdata/colorized/rects.png
